testing/mochitest/tests/Harness_sanity/test_sanityException2.html

<!DOCTYPE html>
<html>
<head>
  <title>Test that uncaught exceptions in mochitests cause failures</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=670817">Mozilla Bug 670817</a>
<script>
SimpleTest.waitForExplicitFinish();
ok(true, "a call to ok");
SimpleTest.executeSoon(function() {
  SimpleTest.expectUncaughtException();
  throw "an uncaught exception";
});
SimpleTest.executeSoon(function() {
  SimpleTest.finish();
});
</script>
</body>
</html>
